+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The IronWeb SDK NPM releases follow standard [Semantic Versioning](https://semve
 
 **Note:** The patch versions of the IronWeb SDK will not be sequential and might jump by multiple numbers between sequential releases.
 
+## v4.3.1
+
+- fix a streaming bug, encrypt could get blocked with no reader yet on writing the header+IV.
+- fix a bug in unmanaged encryption if ironweb was used without a bundler that could provide a `Buffer` polyfill.
+
 ## v4.3.0
 - add streaming encrypt and decrypt functionality for managed and unmanaged documents.